Refactor get started page to welcome and add dashboard as main route
